+++ crates/amaru-ledger/src/rules/transaction/phase_one/mod.rs
@@ -287,7 +287,13 @@ where
         })?;
 
         debug_span!(ledger::rules::phase_one::VOTES).in_scope(|| {
-            voting_procedures::execute(context, era_history, pointer, mem::take(&mut transaction_body.votes))
+            voting_procedures::execute(
+                context,
+                protocol_parameters.protocol_version,
+                era_history,
+                pointer,
+                mem::take(&mut transaction_body.votes),
+            )
         })?;
     } else {
         debug_span!(ledger::rules::phase_one::CERTIFICATES).in_scope(|| {